--- src/components/Constants/icons.jsx
@@ -1,20 +1,0 @@
-import CategoryKoreanIcon from '../../assets/category/category-korean.png';
-import CategoryChineseIcon from '../../assets/category/category-chinese.png';
-import CategoryJapaneseIcon from '../../assets/category/category-japanese.png';
-import CategoryWesternIcon from '../../assets/category/category-western.png';
-import CategoryAsianIcon from '../../assets/category/category-asian.png';
-import CategoryEtcIcon from '../../assets/category/category-etc.png';
-import AddButtonIcon from '../../assets/button/add-button.png';
-
-export const CATEGORY_ICONS = {
-  한식: CategoryKoreanIcon,
-  중식: CategoryChineseIcon,
-  일식: CategoryJapaneseIcon,
-  양식: CategoryWesternIcon,
-  아시안: CategoryAsianIcon,
-  기타: CategoryEtcIcon,
-};
-
-export const BUTTON_ICONS = {
-  ADD: AddButtonIcon,
-};
